pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
  "hatch-vcs>=0.4",
  "hatchling>=1.25",
]

[project]
name = "django-simple-dms"
description = "Django simple Document Management System"
readme = "README.md"
license = { text = "K-Tech" }
license-files = [
  "LICENSE.md",
]
authors = [
  { name = "Giovanni Bronzini", email = "g.bronzini@gmail.com" },
]
requires-python = ">=3.10"
classifiers = [
  "Environment :: Web Environment",
  "Framework :: Django",
  "Framework :: Django :: 5.2",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dynamic = [ "version" ]
dependencies = [
  "django",
  "psycopg[binary]>=3.2.12",
]

[dependency-groups]
dev = [
  "django-environ>=0.12.0",
  "django-extensions>=4.1",
  "httpie>=3.2.4",
  "pre-commit",
]
test = [
  "factory-boy>=3.3.3",
  "freezegun>=1.5.5",
  "pytest>=8.3.3",
  "pytest-cov>=5",
  "pytest-django>=4.11.1",
  "tox>=4.25",
]
docs = [
  "mkdocs>=1.5.3",
  "mkdocs-awesome-pages-plugin>=2.9.3",
  "mkdocs-click>=0.8.1",
  "mkdocs-ezglossary-plugin>=1.6.10",
  "mkdocs-ezlinks-plugin>=0.1.14",
  "mkdocs-gen-files>=0.5",
  "mkdocs-get-deps>=0.2",
  "mkdocs-include-markdown-plugin>=6.2.2",
  "mkdocs-link-marker>=0.1.3",
  "mkdocs-macros-plugin>=1.0.5",
  "mkdocs-material[imaging]>=9.5.15",
  "mkdocs-minify-plugin",
  "mkdocs-redirects",
  "mkdocs-simple-hooks>=0.1.5",
  "mkdocstrings[python]>=0.24.1",
  "pymdown-extensions>=10.7.1",
]
checks = [
  "commitizen>=4.9.1",
  "detect-secrets>=1.5",
  "mypy>=1.17.1",
  "ruff>=0.12.12",
]
[tool.hatch]
build.hooks.vcs.version-file = "src/django_simple_dms/version.py"
version.source = "vcs"
raw-options = { root = "../.." }

#[tool.hatch.build.targets.sdist]
#ignore-vcs = true
#include = [
#  "src/django_simple_dms/**/*.*",
#]

[tool.hatch.build.targets.wheel]
packages = [ "src/django_simple_dms" ]

[tool.uv]
default-groups = [ "dev", "test", "docs" ]
